+++ core/src/main/java/com/radixdlt/p2p/transport/PeerOutboundBootstrapImpl.java
@@ -88,7 +88,7 @@ public final class PeerOutboundBootstrapImpl implements PeerOutboundBootstrap {
   private final P2PConfig config;
   private final Addressing addressing;
   private final Network network;
-  private final String newestForkName;
+  private final ProtocolUpdateId newestProtocolUpdateId;
   private final Metrics metrics;
   private final Serialization serialization;
   private final SecureRandom secureRandom;
@@ -103,6 +103,7 @@ public final class PeerOutboundBootstrapImpl implements PeerOutboundBootstrap {
       P2PConfig config,
       Addressing addressing,
       Network network,
+      ProtocolUpdateId newestProtocolUpdateId,
       Metrics metrics,
       Serialization serialization,
       SecureRandom secureRandom,
@@ -113,7 +114,7 @@ public final class PeerOutboundBootstrapImpl implements PeerOutboundBootstrap {
     this.config = Objects.requireNonNull(config);
     this.addressing = Objects.requireNonNull(addressing);
     this.network = network;
-    this.newestForkName = "SomeForkName";
+    this.newestProtocolUpdateId = newestProtocolUpdateId;
     this.metrics = Objects.requireNonNull(metrics);
     this.serialization = Objects.requireNonNull(serialization);
     this.secureRandom = Objects.requireNonNull(secureRandom);
@@ -137,7 +138,7 @@ public final class PeerOutboundBootstrapImpl implements PeerOutboundBootstrap {
                 config,
                 addressing,
                 network,
-                newestForkName,
+                newestProtocolUpdateId,
                 metrics,
                 serialization,
                 secureRandom,
