+++ core/src/main/java/com/radixdlt/p2p/transport/PeerServerBootstrap.java
@@ -98,7 +98,7 @@ public final class PeerServerBootstrap {
   private final P2PConfig config;
   private final Addressing addressing;
   private final Network network;
-  private final String newestForkName;
+  private final ProtocolUpdateId newestProtocolUpdateId;
   private final Metrics metrics;
   private final Serialization serialization;
   private final SecureRandom secureRandom;
@@ -114,6 +114,7 @@ public final class PeerServerBootstrap {
       P2PConfig config,
       Addressing addressing,
       Network network,
+      ProtocolUpdateId newestProtocolUpdateId,
       Metrics metrics,
       Serialization serialization,
       SecureRandom secureRandom,
@@ -125,7 +126,7 @@ public final class PeerServerBootstrap {
     this.config = Objects.requireNonNull(config);
     this.addressing = Objects.requireNonNull(addressing);
     this.network = network;
-    this.newestForkName = "SomeForkName";
+    this.newestProtocolUpdateId = newestProtocolUpdateId;
     this.metrics = Objects.requireNonNull(metrics);
     this.serialization = Objects.requireNonNull(serialization);
     this.secureRandom = Objects.requireNonNull(secureRandom);
@@ -151,7 +152,7 @@ public final class PeerServerBootstrap {
                 config,
                 addressing,
                 network,
-                newestForkName,
+                newestProtocolUpdateId,
                 metrics,
                 serialization,
                 secureRandom,
