+++ core/src/test/java/com/radixdlt/api/SystemApiTestBase.java
@@ -91,6 +91,7 @@ import com.radixdlt.p2p.P2PConfig;
 import com.radixdlt.p2p.RadixNodeUri;
 import com.radixdlt.p2p.TestP2PModule;
 import com.radixdlt.p2p.addressbook.AddressBook;
+import com.radixdlt.p2p.transport.ProtocolUpdateId;
 import com.radixdlt.rev2.Decimal;
 import com.radixdlt.rev2.modules.REv2StateManagerModule;
 import com.radixdlt.sync.SyncRelayConfig;
@@ -141,6 +142,7 @@ public abstract class SystemApiTestBase {
             new AbstractModule() {
               @Override
               protected void configure() {
+                bind(ProtocolUpdateId.class).toInstance(new ProtocolUpdateId("test"));
                 bind(Network.class).toInstance(Network.INTEGRATIONTESTNET);
                 bind(P2PConfig.class).toInstance(p2pConfig);
                 bind(AddressBook.class).in(Scopes.SINGLETON);
